+++ .github/scripts/.compileLinux.sh
@@ -38,6 +38,7 @@ mkdir ${finalFolder}/Tools
 mkdir ${finalFolder}/Tools/Items
 mkdir ${finalFolder}/Tools/Locale
 mkdir ${finalFolder}/Tools/Themes
+mkdir ${finalFolder}/Tools/OSC
 mv Tools/Items/data* ${finalFolder}/Tools/Items
 mv Tools/Locale ${finalFolder}/Tools
 mv Tools/Themes/*.json ${finalFolder}/Tools/Themes