+++ Globals.py
@@ -1,13 +1,14 @@
 import os.path
 
 # Constants
-_VERSION = 'alpha-0.7.4'
+_VERSION = 'alpha-0.7.5'
 _TITLE = f"ToNCodes {_VERSION} - by MichelleDB"
 _GITHUB = 'https://github.com/69MichelleDB/ToNCodes'
 _WSURL = "localhost"
 _WSPORT = 11398
 _RX = "https://app.tontrack.me/regex.json"
 _DEBUG_REFRESH = 200
+_OSCURL = "localhost"
 
 # Main window
 _WIDTH = 650
@@ -19,15 +20,19 @@ _HEIGHT_MC = 100
 
 # Options window
 _WIDTH_OPT = 610
-_HEIGHT_OPT = 305
+_HEIGHT_OPT = 395
 
 # About window
 _WIDTH_ABOUT = 550
-_HEIGHT_ABOUT = 265
+_HEIGHT_ABOUT = 280
 
 # Debug window
 _WIDTH_DEBUG = 500
 _HEIGHT_DEBUG = 300
+
+# OSC Param window
+_WIDTH_OSCPARAM = 450
+_HEIGHT_OSCPARAM = 500
 
 # Folders
 _FOLDER_CODES = 'Codes'
@@ -38,6 +43,7 @@ _FOLDER_TOOLS = 'Tools'
 _FOLDER_TOOLS_ITEMS = 'Items'
 _FOLDER_TOOLS_THEMES = 'Themes'
 _FOLDER_TOOLS_THEMES_ASSETS = 'Assets'
+_FOLDER_OSC= 'OSC'
 
 # Files
 _FILE_CONFIG = 'config.xml'
@@ -48,6 +54,7 @@ _FILE_DATASILLY = os.path.join(_FOLDER_TOOLS,_FOLDER_TOOLS_ITEMS,'data_silly')
 _FILE_DATAKSILLY = os.path.join(_FOLDER_TOOLS,_FOLDER_TOOLS_ITEMS,'datak_silly')
 _FILE_DATAU = os.path.join(_FOLDER_TOOLS,_FOLDER_TOOLS_ITEMS,'datau')
 _FILE_DATAUK = os.path.join(_FOLDER_TOOLS,_FOLDER_TOOLS_ITEMS,'datauk')
+_FILE_FALLBACKOSCPROFILE = os.path.join(_FOLDER_TEMPLATES,_FOLDER_OSC,'Default.json')
 
 # Global variables
 configList = {}
@@ -69,8 +76,10 @@ regexDict = {}
 codesData = []
 fileBoxSelected = ''
 fileBoxChanged = False
+oscClient = None
+oscJsonProfile = None
 
-# Round variables
+# Round and debugging variables
 roundEvent = ''
 roundNotJoined = -1
 roundMap = ''
@@ -78,6 +87,7 @@ roundType = ''
 roundKiller = ''
 roundCondition = ''
 lastWSMessage = ''
+lastOSCMessage = ''
 
 # TEMP
 auxPathFirstBoot = None