+++ Tools/notesTools.py
@@ -1,7 +1,7 @@
 import Globals as gs
 import re
 from Tools.errorHandler import ErrorLogging
-from Tools.netTools import SendWSMessage
+from Tools.netTools import SendWSMessage, OSCOrder, ExecuteOSCList, SendOSCMessage, ResetRoundOSC, NewRound
 import asyncio
 
 # Check the content of a log and parse all data
@@ -28,6 +28,8 @@ def ParseContent(i_content, i_fileName, i_cursor):
                 key = ''
                 regex = ''
 
+                OSCorderList = []
+
                 for key, regex in gs.regexDict.items():             # Iterate the line with each regex
                     regexMatch = regex.search(lineMatch.groups()[1])
                     if regexMatch:
@@ -52,31 +54,76 @@ def ParseContent(i_content, i_fileName, i_cursor):
                                 ResetRound()
                             case "opt_in":                              # Player joins the game
                                 gs.roundNotJoined = 1
+                                OSCorderList.append(OSCOrder(key, gs.oscJsonProfile[key]['variable'], gs.oscJsonProfile[key]['values']))
                             case "opt_out":                             # Opting out means they respawned
                                 gs.roundNotJoined = -1
                                 gs.roundCondition = 'RESPAWN'
+                                OSCorderList.append(OSCOrder(key, gs.oscJsonProfile[key]['variable'], gs.oscJsonProfile[key]['values'], 'resetRound'))
+                            case "round_start":
+                                OSCorderList.append(OSCOrder(key, gs.oscJsonProfile[key]['variable'], gs.oscJsonProfile[key]['values'], 'newRound'))
                             case "round_map":
                                 gs.roundMap = f"{args[0]} ({args[1]})"
                                 gs.roundType = args[2]
+                                OSCorderList.append(OSCOrder(key, gs.oscJsonProfile[key]['variable'], args[1]))
+                                OSCorderList.append(OSCOrder('round_type', gs.oscJsonProfile['round_type']['variable'], gs.oscJsonProfile['round_type']['values'][args[2]]))
+                            case "round_map_swap":
+                                print('round_map_swap placeholder, I need an example to see how the map is fed, is name and id or just name?') # TODO
                             case "round_killers":
                                 if args[3] != '':       # I found a case with an 8 pages where the game returned an empty round with other killers before the round ended
                                     gs.roundKiller = f"{args[0]} {args[1]} {args[2]}"
                                     gs.roundType = args[3]
+                                    OSCorderList.append(OSCOrder('round_killer1', gs.oscJsonProfile['round_killer1']['variable'], int(args[0])))
+                                    OSCorderList.append(OSCOrder('round_killer2', gs.oscJsonProfile['round_killer2']['variable'], int(args[1])))
+                                    OSCorderList.append(OSCOrder('round_killer3', gs.oscJsonProfile['round_killer3']['variable'], int(args[2])))
+                            case "round_unknown":
+                                print('round_unknown placeholder')
+                            case "round_possessed":
+                                OSCorderList.append(OSCOrder(key, gs.oscJsonProfile[key]['variable'], gs.oscJsonProfile[key]['values']))
                             case "is_gigabyte":
                                 gs.roundType = 'Special'
+                                OSCorderList.append(OSCOrder(key, gs.oscJsonProfile[key]['variable'], gs.oscJsonProfile[key]['values']))
+                            case "is_joy_asleep":
+                                OSCorderList.append(OSCOrder(key, gs.oscJsonProfile[key]['variable'], gs.oscJsonProfile[key]['values']))
+                            case "is_joy_awake":
+                                OSCorderList.append(OSCOrder(key, gs.oscJsonProfile[key]['variable'], gs.oscJsonProfile[key]['values']))
+                            case "is_glorbo":
+                                OSCorderList.append(OSCOrder(key, gs.oscJsonProfile[key]['variable'], gs.oscJsonProfile[key]['values']))
+                            case "is_wild_yet_bloodthirsty_creature":
+                                OSCorderList.append(OSCOrder(key, gs.oscJsonProfile[key]['variable'], gs.oscJsonProfile[key]['values']))
+                            case "is_atrached":
+                                OSCorderList.append(OSCOrder(key, gs.oscJsonProfile[key]['variable'], gs.oscJsonProfile[key]['values']))
+                            case "is_hungry_home_invader":
+                                OSCorderList.append(OSCOrder(key, gs.oscJsonProfile[key]['variable'], gs.oscJsonProfile[key]['values']))
+                            case "is_meatball_man":
+                                OSCorderList.append(OSCOrder(key, gs.oscJsonProfile[key]['variable'], gs.oscJsonProfile[key]['values']))
+                            case "is_foxy":
+                                OSCorderList.append(OSCOrder(key, gs.oscJsonProfile[key]['variable'], gs.oscJsonProfile[key]['values']))
                             case "round_won":
                                 gs.roundCondition = 'WIN'
+                                OSCorderList.append(OSCOrder(key, gs.oscJsonProfile[key]['variable'], gs.oscJsonProfile[key]['values'], 'resetRound'))
                             case "round_lost":
                                 gs.roundCondition = 'LOSE'
+                                OSCorderList.append(OSCOrder(key, gs.oscJsonProfile[key]['variable'], gs.oscJsonProfile[key]['values'], 'resetRound'))
                                 ResetRound()
+                            case "item_pickup":
+                                if args[0] in gs.oscJsonProfile[key]['values']:     # First check if the item exists, that list is not refined yet and new items may enter
+                                    OSCorderList.append(OSCOrder(key, gs.oscJsonProfile[key]['variable'], gs.oscJsonProfile[key]['values'][args[0]]))
+                                else:
+                                    print(f'OSC Alert: item [{rgs[0]}] is not in the json file')
+                            case "item_drop":
+                                OSCorderList.append(OSCOrder(key, gs.oscJsonProfile[key]['variable'], gs.oscJsonProfile[key]['values']))
+
                         
+                        OSCorderList = ExecuteOSCList(OSCorderList)         # Process all the OSC calls
+
                         # Fix name of the round for the Decoder later
                         match gs.roundType:
                             case '8 Pages':
                                 gs.roundType = '8pages'
                             
-                        if key not in ['TONWINTER','TONWAPRIL','TONCODE']:
+                        if key not in ['TONWINTER','TONWAPRIL','TONCODE']:  # Process all Websocket calls
                             SendWSMessage(key, args)
+
                         break
             
         cursor = i_content.tell()                               # If we are done, recover where the cursor is, which should be at the end
@@ -84,7 +131,8 @@ def ParseContent(i_content, i_fileName, i_cursor):
     except Exception as e:
         print(e)
         ErrorLogging(f"Error in ParseContent: {e}")
-
+    
+    
 # Reset all Round variables
 def ResetRound():
     gs.roundMap = ''